Drive and Steer logging
I added AKit logging to the drive modules as well. This will list the current draw for all 8 swerve drive motors in AKit, labeled by which motor is which.

Definitely review my work on this one, I let AI figure out how to do this, since integrating with the 6328 drive code is more difficult.