More climber logging
I added more logging to the climber subsystem. I am trying to capture commanded states, outputs, and any flags we use for reference elsewhere in the code. This subsystem is a little harder for me to follow the flow of, so I'm sure there are still major gaps in the data here.