Initial import
I imported the class that logs these preferences. I also added a few lines to trigger this logging when we enter auto and when we enter teleop. This should "just work" but I haven't tested this on any hardware yet.

This is a branch off the summer branch so we can easily test it on the summer robot as it stands.